Improving handling of different product data types/new products
- When generating cogs data are read based on their data type
- Improvements to output cogs (generated files should be smaller)
- Added Dry Matter Productivity to system

+++ StatsExtractorCpp/StatsExtractor/lib/Utils/Utils.cpp
@@ -146,8 +146,7 @@ std::string rgbToArrayString(RGBVal &array) {
     std::stringstream k;
     k <<"[";
     for(auto& it:array)
-        k << it <<",";
-
+        k << static_cast<unsigned int>(it) <<",";
     k.seekp(-1, std::ios_base::end);
     k <<"]";
     return k.str();
